+++ AreaOfaPoligon/src/ru/example/Position.java
@@ -1,16 +1,16 @@
 package ru.example;
 
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private boolean wheather;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public Position(int x, int y, boolean wheather) {
+    public Position(double x, double y, boolean wheather) {
         this.wheather = wheather;
         this.x = x;
         this.y = y;
@@ -21,19 +21,19 @@ public class Position {
         return "x: " + x + " y: " + y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
